Add composable-forms and rip out elm-validate
This begins a major refactor of the entire Editor - moving away from a
lot of boilerplate hopefully! I'm hopeful that using composable-forms
will make the code more consise and simple, and cutting down on a lot of
extra, custom code. Of course, I've learnt a lot from doing it the more
laborious, manual way - a file with around 800 lines! Now I'm hoping
that redoing it will make the end product so much better. And that would
not necessarily have been the case If hadn't gone done the other route
at first!