Test etaque/elm-form
Latest challenge was getting adding tag inputs onEnter working. Seems
quite promising!
The library is quite a bit easier to work with as opposed to
composable-form, however it's also a bit more limiting. But I'm getting
there. Some of the behaviour that I want - sending a default value for a
new entry in a list of entries (e.g. after hitting enter on a tag input
field) requires a bit of a hack. However, it does seem to work.