Extract form from Test into own module
It does make the Test page much more light-weight. The idea is to now
simply use the new Form module in the Editor module. Letting the Editor
code handle server communication etc and all the stuff around the actual
form. And the actual form will handle the form.

One downside of this is that I must (?) use the translator pattern,
mapping subModels and subMsgs back and forth etc in the Test page. It's
not simply a matter of outsourcing the code just like that...