Make more progres on using the Form in the Editor
There is the small matter of handling form submit. I will probably need
to use the Translator pattern by transforming Form.Submit in Form to
something like SubmitValidForm Encode.Value which the parent (Editor)
can pick up and POST to the server etc. However, doing so is not
trivial, I will need to pass it through the runtime and it is a bit of a
hack. References:
- https://medium.com/elm-shorts/how-to-turn-a-msg-into-a-cmd-msg-in-elm-5dd095175d84