Use the new etaque/elm-form Form in the Editor fully
It now supports all features the old one did. The form resides in
its own module. However, it is a bit of a leaky abstraction (all
abstractions are though), since the Editor need to map child Msg:s and
specifically need to deal with the "SubmitValidForm Encode.Value" that
the child (Form) emits when the user submits the form after succesful
validation. This modularisation incurs some extra overhead that could've
been skipped if everything was kept in the same module...