Do not double %-encode slug - fixing 406 issues
It should now be fine to have whatever title you want (with any amount
of reserved characters in it) and still be able to navigate to it
without issue.
The "slug" is actually just the title of the recipe - so it is not
beautified or anything. However, Firefox automtically displays the
decoded variant - producing quite good looking urls (with %20 shown as
spaces - but turned into %20 when copied).
Chrome doesn't do that.