Add prefix matching to search
Still need to fix the construction of the tags tsvector and include
ingredients tsvector in the search. Currently I can build an ingredients
tsvector but it's not very dry and it's done in the view - so no
index...

The search is implemented as a stored function, exposed at /rpc/search
and takes a json document with key "search_query" with its value being
the actual search query. That is then converted into a tsvector for
stemming, then each term is suffixed with ':*' to implement prefix
matching. All the terms in the query are added together with '&' to
construct a proper tsquery that will match partial (prefix) words.

Ref: https://stackoverflow.com/questions/46122175/fulltext-search-combined-with-fuzzysearch-in-postgresql