Refactor db model - ingredients in a jsonb col
Abondoning the relational model to store ingredients has some benefits.
For full text search we can have a single generated column on recipes
that we can index, spanning both metadata on the actual recipe and
ingredients. This allows for a more smooth implementation of search and
it should also be faster.
Secondly, I didn't use any strengths that the relational could have been
used for. It was simply transformed into json when read, and input was
transformed to the relational model before being written. This was a
waste of CPU cycles and developer complexity.