Add new endpoint /images/upload for direct binary upload
It accepts Content-type: image/* direct binary image uploads, reads
it in memory (does not buffer into temporary file - altough it's trivial
to implement, but I see no reason why at this point in time), then
calculates a random hash and returns it in the JSON response.

The idea is that the client will asynchronously upload images with POST
requests to /images/upload, which will return the new url for the saved
image. Then the client will include this url in the POST/PATCH requests
to the recipes endpoint, something like:
{
  title: "Cheese Cake",
  ingredients: "...",
  header_image: { url: "/images/xyz.jpeg" },
  ...
}
The PostGREST endpoint /recipes will then save that url, and serve it on
subsequent requests for the recipe.

Of course, a number of issues arise:
- the client will need to wait for the response from /images/upload
  before POST:ing the recipe
- the clients could upload a bunch of files, then exit the session, does
  polluting the /uploads directory with a bunch of never to be used
  files. I need to write a script that cleans up this directory
  periodically.

TODO:
- [ ] mount a volume /uploads in openresty
- [ ] endpoint to actually serve the images - with image processing
  capabilities
- [ ] periodically clean up unused uploads (older than x hours)
- [ ] accept images in the main recipe model (e.g. header_image)