💄 Add Metropolis font
Metropolis has 9 different weights, but they are named slightly
differently in elm-ui, and needs to be referred by numbers in css:

font-weight   | Metropolis  | elm-ui font weight
------------- | ----------- | -----------------
100           | Thin        | harline
200           | Extra Light | extraLight
300           | Light       | light
400           | Regular     | regular
500           | Medium      | medium
600           | Semi Bold   | semiBold
700           | Bold        | bold
800           | Extra Bold  | extraBold
900           | Black       | black

For every font weight, there are two styles, one normal and one italic.

I couldn't find a CDN that hosts this font, so had to download the
source and include it a custom @fontface css in fonts.css. This file
could/should be inlined. Also, it's possible to include woff version 1
files as fallback as well, but it seems woff2 has pretty good support
so I didn't do that.