Use dillonkearns/elm-markdown to render markdown
The big idea is to use elm-markdown to render the markdown, giving me
the ability to actually render it to elm-ui, and not through Html, as
was the case previously. Furthermore, it supports custom html tags, so I
can support a <timer name="eggs" seconds=360> element that I have full
control over to render with elm-ui! Very promising!

TODO:
- lots of things in terms of tweaking the presentation, spacing,
  padding, font size, font weight etc
- make task lists clickable and have that mean something - so don't
  simply use Input.defaultCheckbox
- <timer> custom Html tags
- make a Validator that uses elm-markdown, I can restrict to a subset of
  valid markdown with that, e.g. only allow unordered lists and H2:s in
  ingredients input