✅🐞 Add tests for image server, fix bug in image server
Suprisingly smooth to test this with js - even though it involves binary
files being uploaded and downloaded!
But it feels really worthwhile to have these tests, I already discovered
and fixed a bug due to them!
The 🐞: crashing if uploading image without specifying content-type.

TODO:
- clear up the cache after each test run
- refactor image_server tests into own sub dir - I was afraid to do
so at first due to all the stuff that looked a bit scary in common.js,
but now I'm not afraid of that anymore!
- test the cryptographic signatures, and actually implement them!