👷‍♂️ Build frontend in its own container
It's nice, because it means that I don't bust the openresty cache every
time I need to rebuild the frontend (which is probably almost on every
deploy), and it means a better separation of concern, the
frontend/Dockerfile is the one that builds the frontend, not the
openresty/Dockerfile.

It works by building the frontend/Dockerfile before the openresty
service as "frontend_builder" in docker-compose.yml, then copying from
that image the build artifacts, into the openresty at end of that
image building.