🐞 Respond with application/json if error in @image_server
The internal nginx @image_server location should only respond to client
directly when it errors out, in which case we don't want the clients
requested file extension to determine the content-type, but we want json
in an error message!
In the success case, @image_server redirects to the same location again,
this time try_files should find the file in the cache, and then use the
mime.types to determine the content-type header from the file extension!

Add eslintrc.js symlink in root dir as well, for the tests.