⚡️ Sprinkle some lazy to improve performance
Quite specualtive how much this will improve things, and the second part
is Keyed, unfortunately, it is not possible to use Keyed with wrappedRow
in elm-ui. And, the other part where Keyed might make sense is in
ingredient list, but that is rendered with markdown, and I currently
feel that is overkill, and not worth the effort.
Todo: actually evaluate the performance impact if these changes.