💩 wip: Try to implement a smooth transition from blur
Might be an issue with timing with removing image class - perhaps
ideally it should not be removed until the new img has actually been
loading?

Some thoughts that might be useful here:
https://www.sitepoint.com/how-to-build-your-own-progressive-image-loader/

Also, there is a native loading="lazy" that can now be added to html
image elements! Unfortunately annoying to add in elm-ui because the img
tag is wrapped in a div, which all my attributes stick to instead...