Create cookie session on /login endpoint w. openresty
Example:
http --session=./session.json POST :8080/login email="alice@email.com"
password="pass"

You will get back the "me" key in the json from PostgREST, but the jwt
token will have been stripped and added to a session cookie instead.
With the --session to httpie, it will save that into a session.json
file, and you can continue query the API with the session cookie in
place.

Next step is to protect all routes in OpenResty by requiring the session
cookie to have been set. And, pass on the jwt token into a header to
PostgREST from the cookie.

Then I can disallow the db to show any data with the anonymous role.