Protect /rest endpoints with access_by_lua_block
The block checks for a valid session cookie, if found it extracts the
jwt, sets it into the "Authorization: Bearer <token>" header, that
PostgREST then uses to authenticate the user.

For example:
http :8080/rest/recipes will return a 403 because we do not supply a
valid cookie, but:
http --session=./session.json :8080/rest/recipes will succeed (provided
that we have done a http --session=./session.json /login beforehand
successfully).