Protect image_server with session cookie auth
Now the tests are also updated and should provide a fairly reasonable
coverage.
Also revoked most rights for the "anonymous" user in the database, will
need to manually do these steps upon deploy.
As well as to change jwt_lifetime: it needds to be set longer than the
cookie expiration time in this setup.