Clean up Recipe.ServerError
There is no point in decoding to "PgError" or whatever, I only ever care
about seeing it as a string (and even that should ideally never be
displayed to the user), and the default string representation from the
http response (a json value if the server is somewhat healthy) is good
enough.
Of course, I would ideally like to pretty print it, but then I do have
to parse it, and then pretty print, and it is just not worth it.

A notable improvement with this is that I have made the Recipe request
interface more homogenous by always returning Result ServerError and not
Http.Error for Delete only...