🗃  Add db migrations for auth/cookie session management
Something is not quite working as intended though...
When pulling prod db, migrating, and logging in through ui or
/rest/login via cli, I get a cookie back as expected. But, when then
attempting to use it (via either cli or ui) I get an error:

In json body: { message: "JWSError JWSInvalidSignature"}
and in header:
www-authenticate: Bearer error="invalid_token",
error_description="JWSError JWSInvalidSignature"

I've seen this before, and I am not sure what is the issue.
Possibly something with the secret?

**UPDATE**
The fact of the matter is, you need to have the same jwt_secret in both
postgrest and postgres, duh. So both `- JWT_SECRET=${JWT_SECRET}` for db
and `PGRST_JWT_SECRET=${JWT_SECRET}` for postgrest need to match, which
they do when starting from scratch.
But, when imporing the prod db, the secret is encoded into the database
table 'select settings.get('jwt_secret', 'xxx')', and won't change
simply because of an env var.
So, the simple solution is to set your local .env to match prod's .env,
or (which might be preferable), to set the secret in the local db.

TODO:
- document migration flow (sqitch, scripts/sqitch, subzero migrations)
- figure out if one should use sqitch locally against another target db,
  since subzero migrations' apgdiff is sort of in conflict with the
  normal subzero dashboard - when deploying, it will complain that
  certain db objects already exists, because the target doesn't have a
  sqitch schema, it believes it is empty, but is not, it contains the
  seeded data from init.sql...