⏪ Disable progressive blurhash image loading
It is not quite ready for prime time just yet.
When it works, it is actually pretty nice, the only things missing are:
- individual blurhashes (needs to be pre-computed for each image, so
  that the server actually responds with it)
- animate the transition out when the blurred image is replaced with the
  unblurred variant: currently the full-res image is nicely transitioned
  in with a css transition directive (blur);

  ```css
  .blurred img {
    filter: blur(0.2em);
  }

  img {
   filter: blur(0em);
   transition: filter 0.5s;
  }
  ```

  but, the blurHash should not immediately be removed from the DOM
  ideally, but nicely transitioned out first. This is tricky with my
  current setup, (it is a classic challenge with animations I do
  believe). Potentially elm-animator will make it a bit nicer to do,
  eventually...

The big blocker now though is that sometimes it won't load images at all
the blurhash is displayed forever, until I click arond, refresh, click
around some more, then refresh, or something... Not a great UX
experience. So this the absolute ⛔!