💄💩  Play around with progressive image loading
I am now back to using Background.image which appears to give me a few
benefits:
- it works as intended layout wise on all devices afaict
- it is pure Elm ui, no css object-fit: cover or somesuch nonsense
However, if I want to blur it, I can only blur the entire div, including
the title text...
There is a very promising css property, backdrop-filter, but appears to
only have limited support, and even in Chrome (where it is supposed to
work), it does not seem to work correctly when I apply it, though the
sample on css-tricks works, and Chrome devtools recognise the attribute,
it just doesn't seem to want to blur my images with it...

Also, still having issues with the images never becoming fully loaded at
times - especially in Chrome, I have noticed that a Cmd-Shift-R works,
so figuring it *might* have something to do with the Service worker
caching stuff or something like that...
**Update**: nope, it is not sw, this appears on dev server (where no sw
is running), but, likely the issues is the setTimeout of 0 was not
enough, changed it to 10 makes it work well, afaict.

Anyway, I wanted to have a branch with these experimental changes, so
that I can decide if I will pursue this further or not.

In conlusion, there are three reasons that I will not continue down this
avenue any further:
- the bugginess that sometimes it doesnt't appear to work at all, likely
  fixed by setTimeout 10, but still feels wrong
- the choppy transition animation - only animating blur(0.2em) ->
  blur(0) is not 💯, ideally I will want to first fade out the blurHash
  image, however this is quite a large project
- with Background.image I cannot blur only the background image, (with
  backdrop-filter as mentioned above I should be able to, but browser
  support seems spotty), and I don't want to use an image element as it
  is hacky with Elm + CSS and the hack I used did not work on iOS...

So issue 1 is probably fixable, issue 2 can be fixed but require a lot
of effort, issue 3 might be fixable, but I really don't know for sure.
Considering all of this, I think it makes sense to focus my attention on
other issues which will have a larger positive overall impact on the
app. Who knows, maybe one day I'll return to to this branch though.