🔥➕ Add elm-review with default no unused conf, and remove deade code
I kept SendPortMsg custom type constructor for now, event though it is
not used, as I felt for doc purposes it might make sense to keep, should
I wish to send some message to the editor. Although, if so I would need
to actually do something on the js side of things...
I might remove it in the next commit - at least make things easier to
revert and find in git should I wish to go back.