💄 Always render unordered lists in instructions as task lists
...but not in ingredients list.
The downside: it is now not possible to have e.g. a list of notes in
instructions without it turning into a task list.
I think the trade off is worth it currently. The alternative is to have
better recipe source markdown, but that is more manual work. And,
ideally should be accompanied by an easier MDE interface for writing
task lists - something I find rather fiddly actually.
And, then I would need to add a button for that too I think (or make the
default list button be a task list button).