🐛 Fix multiple task lists sharing same index bug
Turns out I didn't have to use randonmness to generate a key, but could
use the excellent elm-markdown package to add "metadata" about which
index each markdown list has - and use that to give a unique index to
each list.
TODO: extract some logic that now resides in the render with a
transformation of the actual markdown AST, which is super convenient
with elm-markdown. E.g. whether to always treat all unordered lists as
TaskLists or not should not be determined in the rendering function. If
there is a TaskList - it should be a task list, if not, it shouldn't be,
the transformation should happen on the AST instead. Super clean!

This fixes #35