✨ wip: Add Ingredient module to parse and scale ingredient strings
The idea is that this module can be used to scale the number of
portions, by parsing strings of ingredients into Ingredient, which then
can be scaled, then toString:ed to get the new result.

The question is how to parse these ingredient strings, how to make that
robust. I quite like elm/parser, but the question is if it is overkill?
Also, how to do error handling? I am thinking now that it would make
sense to have a graceful fallback to not scale such invalid strings at
all, just to let them be.
However, one could go different routes, for example:
- be super strict and mandate that all ingrdient definitions follow the
  same specific pattern of "quantity unit ingredient text", and could
  try and enforce this when uploading too, and if error,
  Ingredient.fromString would not be a total function so to speak, but
  would return an Err. Which could perhaps at that point be gracefully
  mapped back to the original string, simply (possibly with a UI
  indicator that I couldn't scale this specific ingredient)
- be very lax, scale the ones I can, not the rest, also to make the
  Ingredient type very flexible, Ingredient.fromString should be total,
  which could be possible with something like `{quantity: Maybe String,
  unit : Maybe String, ingredient: String}`, and worst case put
  everything into `ingredient`. This actually might be a good idea...
  In any case I think it makes sense to keep this type opaque. The
  question is if the module should return `Result`, I think that might
  be that case irrespective if I chose Regex or Parser approach.