Improve API of Markdown module
It no longer relies on Ingrdient - but receives a mapping function that
transforms all the taskItems in the input markdown. Of course this
function is custom made to be used by the Page.Recipe module and to
apply the scaling.
However, I still believe this separation makes more sense - now the
dependency on Ingredient is in Page.Recipe and not in Page.Recipe.Markdown!