Run frontend/Dockerfile on amd64 explicity
Because I currently use create-elm-app, and I do it in docker, in the
"frontend_builder" step as used in deploy.sh and implicitly by the
building of openresty/Dockerfile, *and* create-elm-app has a dependency
on the *npm* package "elm", which is *not* packagad for aarch64/amd64, I
need to use emulation as I do here.
Alternatively, as on-metal, elm and create-elm-app works flawlessly
afaict through Nix, I should also build the frontend w. Nix. But, that
is another beast to tackle another time - and probably I should switch
out create-elm-app (because unmaintained).