🔥 Remove node2nix
For now, little point in also building the frontend with Nix.
The point might be to get rid of frontend/Dockerfile which is used as
"frontend_builder" in the deploy script and in openresty/Dockerfile
(COPY --from frontend_builder /dist/...). But better to keep all of prod
in docker compose until I can deploy it differently anyway.