🐛 Fix usage of lua_resty_session
As I updated the lib to a later version the API significantly changes,
which caused all sorts of buggy behaviour.
With this commit I now hopefully use the lib correctly, though there is
a bit of a quirk in that it sets two cookies: one persistent and one
session. And I can't seem to get rid of the session one, but no harm no
foul I think...