🐛 Don't immediately switch to recipelist, but go through login form msg
The working theory around the bug: it's Chrome trying to autofill and it
does so with an undefined JS value, which makes it uncatchable (ish) in
Elm-land. Due to timing differences on prod, it only manifests there and
not locally: does autofills next field available which happens to be
search box.

Prevention:
- some way to make Chrome not do that? (unclear how though)
- fiddle with code path as I do here: maybe will make it trigger through
the input fields somehow
- maybe setting correct id for search box was enough?