🐛 Attempt to fix duplicate markdown view output
v7.0.1 of dillonkearns/elm-markdown
mapAndAccumulate returns the mapped result and then the original (in certain cases) -> causing duplicate output
https://github.com/dillonkearns/elm-markdown/issues/117
only do this in certain cases... wtf?
solution: just keep the first part...
problem: not really a solution because sometimes we lose contents with this...
E.g. seems to be functioning if input is Ordered/Unorderedlist but not
otherwise?