➖ Remove direct dep on elm from package.json in frontend
The npm package elm for the compiler is not recommended, and I don't use
it locally as I use the one from Nix (which is actually packaged for
apple silicon btw).
But unfortunately parcel's elm integration still relies on the elm npm
package, so it remains in my indirect deps.