🏗 Have /export_to_list expect json payload with ingredients
Expecting body with {"ingredients": [<list of ingredients>]}
allows the frontend to, in the future, not simply add all ingredients of
the recipe, but selectively choose which ones to add.
It also move ingredient markdown parsing responsibility to the frontend
where we have 1) Elm and 2) elm-markdown to help, instead of Lua (though
that also worked fine in practice) on the BE, and 3) already a
responsibility around handling the semi-structured ingredient markdown.