🚚 Extract db service into db/service.nix
Keeping the top-level flake.nix very clean indeed.
Question is, even if openresty-receptdb is technically not a service
(it's just a settings.process - a "command"), why not extract that too?
But then what about postgrest?
Or just keep it a bit more in flake.nix and not modularise too much?