✨ Dynamically fill in rich link preview template by recipe title
**Open graph meta tags and twitter:card tags template into index.html**
The frontend/src/index.html is now treated as a lua-resty-template html
template. So the values in {{ x }} is interpolated when rendering in
openresty lua code. Values in {* x *} are also interpolated but as lua
expresssions - meaning no HTML escaping is done here (for the url:s).
Since we us Parcel as a bundler and it does its own preprocessing we
need to escape those template expressions, this is set in parcelIgnore
in frontend/package.json.

**Openresty rich link preview with model public_recipes **
Openresty adds a route for /recipe/title to not immediately just send
back the index.html with the SPA that then would authenticate with JS
and then call the REST API, but instead enhances - renders the template
with the appropriate values. It does so by fetching from a new view
model that is publice accessible - public_recipes (anonymous user in
Postgres/t can access it). So there's a new location in nginx for
accessing this: /internal/public/ - which does *not* pass in an auth
header.

Similarly to this there's a public image server that in contrast to the
one behind auth does not check the signatuer, and only allows slightly
smaller images.