🔺 Troubleshoot image serving...
Something is not right.
It seems the public image server does everything right - it resizes and
saves valid images in the uploads cache, but when it comes time to serve
it, it simply doesn't work as intended. The response looks fine from
curl, and some sort of bytes are saved locally, but won't open as an
image...