🔒 Verify all image endpoints
The api.recipes jsonb is now enriched with url1496 and url1600 that
contains looks like: /images/<sig>/1496/file_name.jpeg or
/images/<sig>/1600/file_name.jpeg, respectively. Where the <sig> is a
valid hmac signature over the size and file name of the image.
Also slight refactor  - rename to resize_image.lua - as that's all it
does, it doesn't actually serve the image.

The images API should now be even more robust against a DOS attack
trying out thousands of various resize requests - the only allowed ones
will be 1600 and 1496 for now.

Requires to add the env var $IMAGE_SERVER_SECRET to both the db and
openresty.