🚑 Fix utils.get_base_url()
The issue is that running behind the nginx proxy - we pick up the
virtual_port being 8081 and the lua utils.get_base_url() functions
returns https://recept.axellarsson.nu:8081 - which is obviously wrong.
This fixes that - if behind proxy - never set the port - assume we serve
on 443.

As a fallback respect SITE_BASE_URL though.