🐛 Fix url slug bug
Certain recipes containing `&` would not load.
The issue was in constructing the href link from the main recipe listing
page, it simply did e.g. `href="/recipe/Små goda bullar` - taking the
slug as is (as the String title of the recipe). That's actually an
invliad url but when then actually clicking the link the browser tries
to construct a valid url so it url-encodes it, *but*: it can't
reasonabley url-encode the `&` - because that could be meaningful, so it
constructs: `/recipe/Sm%C3%A5%20goda%20bullar` which is fine and works,
however, if there is an ampersand, `&` in there it would be retained,
e.g.
`/recipe/Lammracks%20med%20potatis%20&%20f%C3%A4nk%C3%A5lsgrat%C3%A4ng`
Note the `&`                         ⬆
Then, when fetching the recipe from the URL, the SPA decodes the URL
into a title, from the slug and tries to find a recipe from the server,
however, the url constructed becomes messed up if the `&` is not
url/percent-encoded!

The solution is of course to percentEncode valid urls already in the
listing, then the rest works as intended.

Another option, would be to construct a proper slug that is not simply
the title, but e.g. kebab-case with the `&` stripped out:
`lammracks-med-potatis-fankalsgratang`. But that requires some server
side changes and other considerations. So future improvement!