📌 Pin webauthn to 2.5 in db/Dockerfile to trigger db image build
A bug on prod with passkeys prevented it from functioning - turns out
haven't rebuilt the db container for a year due to no changes being done
in db/ dir - so the GH Actions wasn't triggered. Of course, locally I
build from source tree and thus got a newer webauthn version (2.5.2),
while prod is stuck on 1.11.1 or something! Naturally that broke things
as I upgrade client code for 2.5.

This pins the db webauthn to ~=2.5 which:
- should trigger a rebuild in GH actions
- should fix the passkey errors as we upgrade webauthn lib to 2.5.2